Finals changes in Sergey-bank-project
