+++ Sergey_tasks/JS/Bank-project/src/Bank.html
@@ -7,8 +7,8 @@
       <link rel="shortcut icon" type="image/png"/>
       <link href="https://fonts.googleapis.com/css?family=Poppins:400,500,600&display=swap"
       rel="stylesheet"/>
-      <link rel="stylesheet" href="Bank.css" />
-      <title>Bankist</title>
+      <link rel="stylesheet" href="Bank.css"/>
+      <title>Bankist-new</title>
    </head>
    <body>
     <!-- TOP NAVIGATION -->
@@ -97,14 +97,14 @@
 
          <!-- LOGOUT TIMER -->
          <p class="logout-timer">You will be logged out in 
-            <span class="minutes">00</span>:<span class="seconds">00</span>
+            <span class="minutes"></span>:<span class="seconds"></span>
          </p>
       </main>
 
       <!-- <footer>
       &copy; by Jonas Schmedtmann. Don't claim as your own :)
       </footer> -->
-
+      <script src="const.js"></script>
       <script src="Bank.js"></script>
    </body>
 </html>
