+++ Sergey_tasks/JS/Bank-project/src/Bank.scss
@@ -18,7 +18,7 @@
  body {
    font-family: "Poppins", sans-serif;
    color: #444;
-   background-color: #f3f3f3;
+   background-color: #dfee90;
    height: 100vh;
    padding: 2rem;
  }
@@ -130,22 +130,12 @@
    overflow: scroll;
  }
  
- .movements__row-deposit{
+ .movements__row{
    padding: 2.25rem 4rem;
    display: flex;
    align-items: center;
    border-bottom: 1px solid #eee;
-  //  display: none;
-  // opacity: 0;
  }
- .movements__row-withdrawal{
-  padding: 2.25rem 4rem;
-  display: flex;
-  align-items: center;
-  border-bottom: 1px solid #eee;
-  // display: none;
-  // opacity: 0;
-}
 
  
  .movements__type {
