+++ Sergey_tasks/JS/Bank-project/src/const.js
@@ -1,0 +1,32 @@
+
+const labelWelcome = document.querySelector('.welcome');
+const labelDate = document.querySelector('.date');
+const labelBalance = document.querySelector('.balance__value');
+const labelSumIn = document.querySelector('.summary__value--in');
+const labelSumOut = document.querySelector('.summary__value--out');
+const labelSumInterest = document.querySelector('.summary__value--interest');
+const labelTimer = document.querySelector('.timer');
+//---------------------------------------------------------------------------
+const containerApp = document.querySelector('.app');
+const containerMovements = document.querySelector('.movements');
+const movementsRowDeposit = document.querySelector(".movements__row");
+const movementsTypeDeposit = document.querySelector(".movements__type-deposit");
+const movementsTypeWithdrawal = document.querySelector(".movements__type-withdrawal");
+const movementsValueDeposit = document.querySelector(".movements__value-deposit");
+const movementsValueWithdrawal = document.querySelector(".movements__value-withdrawal");
+//---------------------------------------------------------------------------
+const btnLogin = document.querySelector('.login__btn');
+const btnTransfer = document.querySelector('.form__btn--transfer');
+const btnLoan = document.querySelector('.form__btn--loan');
+const btnClose = document.querySelector('.form__btn--close');
+const btnSort = document.querySelector('.btn--sort');
+//---------------------------------------------------------------------------
+const inputLoginUsername = document.querySelector('.login__input--user');
+const inputLoginPin = document.querySelector('.login__input--pin');
+const inputTransferTo = document.querySelector('.form__input--to');
+const inputTransferAmount = document.querySelector('.form__input--amount');
+const inputLoanAmount = document.querySelector('.form__input--loan-amount');
+const inputCloseUserName = document.querySelector('.form__input--user');
+const inputClosePin = document.querySelector('.form__input--pin');
+
+
